Russian APT/APT28-Adversary-Simulation/requirements.txt
pyautogui
pyvirtualdisplay
requests
cryptography
